+++ include/mpp/algo/lu.hpp
@@ -19,11 +19,11 @@
 
 #pragma once
 
-#include <mpp/detail/utility/algorithm_helpers.hpp>
+#include <mpp/detail/util/algo_impl.hpp>
 #include <mpp/detail/utility/cpo_base.hpp>
-#include <mpp/detail/utility/utility.hpp>
+#include <mpp/detail/util/util.hpp>
 #include <mpp/mat/matfwd.hpp>
-#include <mpp/utility/sq.hpp>
+#include <mpp/util/sq.hpp>
 
 #include <cassert>
 #include <cstddef>
