+++ include/mpp/arithmetic/divide.hpp
@@ -21,7 +21,7 @@
 
 #include <mpp/detail/expr/expr_binary_op.hpp>
 #include <mpp/detail/expr/expr_binary_val_op.hpp>
-#include <mpp/detail/utility/algorithm_helpers.hpp>
+#include <mpp/detail/util/algo_impl.hpp>
 #include <mpp/mat.hpp>
 
 #include <algorithm>
