+++ include/mpp/detail/util/algo_impl.hpp
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include <mpp/detail/utility/utility.hpp>
+#include <mpp/detail/util/util.hpp>
 #include <mpp/mat/matfwd.hpp>
 
 #include <cassert>
