+++ include/mpp/util/sq.hpp
