+++ include/mpp/util/traits.hpp
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include <mpp/detail/utility/utility.hpp>
+#include <mpp/detail/util/util.hpp>
 
 #include <complex>
